+++ internal/controller/operator/factory/vtcluster/select.go
@@ -247,7 +247,7 @@ func buildVTSelectPodSpec(cr *vmv1.VTCluster) (*corev1.PodTemplateSpec, error) {
 	storageNodeFlag := build.NewFlag("-storageNode", "")
 	storageNodeIds := cr.AvailableStorageNodeIDs(vmv1beta1.ClusterComponentSelect)
 	for idx, i := range storageNodeIds {
-		storageNodeFlag.Add(build.PodDNSAddress(cr.PrefixedName(vmv1beta1.ClusterComponentStorage), i, cr.Namespace, cr.Spec.Storage.Port, cr.Spec.ClusterDomainName), idx)
+		storageNodeFlag.Add(vmv1beta1.PodDNSAddress(cr.PrefixedName(vmv1beta1.ClusterComponentStorage), i, cr.Namespace, cr.Spec.Storage.Port, cr.Spec.ClusterDomainName), idx)
 	}
 	if len(cr.Spec.Select.ExtraStorageNodes) > 0 {
 		for i, node := range cr.Spec.Select.ExtraStorageNodes {